Fix server-endpoint related to getting users
Previous behavior had tests, but they did not represent the current state of how the tested-endpoint behaves.
New behavior provides tests which call the actual endpoint to get the data to be used in the tests.
